requirements.txt
streamlit
opencv-python
pillow
torch
numpy<2.0.0
ultralytics